+++ src/main/resources/mapper/project-mapper.xml
@@ -29,6 +29,10 @@
         where project_no = #{project_no};
     </select>
 
+    <select id="isProjectExistById" parameterType="int" resultType="boolean">
+        SELECT EXISTS(SELECT * FROM project_table WHERE project_no = #{project_no});
+    </select>
+
     <!--  참가자 부분          -->
 
     <select id="getParticipants" parameterType="int" resultType="ProjectParticipants">
@@ -47,10 +51,11 @@
         values (#{projectNo}, #{participantNo}, #{participantRole})
     </insert>
 
-    <select id="getParticipant" resultType="ProjectParticipants">
-        SELECT *
-        FROM project_participants
-        WHERE project_no = #{projectNo} and participant_no = #{participantNo}
+    <select id="isParticipant" resultType="boolean">
+        SELECT EXISTS(SELECT *
+                      FROM project_participants
+                      WHERE project_no = #{projectNo}
+                        and participant_no = #{participantNo});
     </select>
 
     <!-- category | role 부분-->
